requirements.txt
numpy<2.0.0
pymupdf==1.23.14
sentence-transformers==2.2.2
chromadb==0.4.18
rank-bm25==0.2.2
spacy==3.7.2
pyyaml==6.0.1
tqdm==4.66.1
tiktoken==0.5.2
pydantic==2.5.0
google-genai>=0.3.0
